Add: CreateUserのメソッドを追加
