+++ containers/api/internal/core/mock_gateway.go
@@ -1,27 +1,40 @@
 package core
 
 import (
+	"context"
 	"sync"
 )
 
-// MockGateway TODO
+// MockGateway - MockDBのアダプターの構造体
 type MockGateway struct {
 	db *MockDB
 }
 
-// MockDB TODO
+// MockDB - テスト・開発用のDB
 type MockDB struct {
 	mu    sync.RWMutex
 	data  map[uint]User
 	index uint
 }
 
-// NewMockDB TODO
+// NewMockDB - テスト・開発用のDBのコンストラクタ
 func NewMockDB() *MockDB {
 	return &MockDB{data: make(map[uint]User)}
 }
 
-// NewMockGateway TODO
+// NewMockGateway - MockDBのアダプターの構造体のコンストラクタ
 func NewMockGateway(db *MockDB) Repository {
 	return &MockGateway{db}
+}
+
+// CreateUser - Userを登録
+func (r *MockGateway) CreateUser(ctx context.Context, user User) (User, error) {
+	r.db.mu.Lock()
+	defer r.db.mu.Unlock()
+	r.db.index++
+
+	user.ID = r.db.index
+	r.db.data[user.ID] = user
+
+	return user, nil
 }
