+++ containers/api/internal/core/validate.go
@@ -1,4 +1,4 @@
-package user
+package core
 
 import "github.com/go-playground/validator/v10"
