--- containers/api/internal/user/repository.go
@@ -1,8 +1,0 @@
-package user
-
-import "context"
-
-// Repository TODO
-type Repository interface {
-	CreateUser(ctx context.Context, user User) (User, error)
-}
